Update Go Modules and Refactor Broker Logic for Enhanced Functionality
- Added `github.com/phuslu/log` as a direct dependency in `go.mod`.
- Removed unused indirect dependencies from `go.mod` and `go.sum`.
- Refactored `Balance` and `Desk` components to improve data handling and management of positions.
- Introduced new methods in `Position` for better order and execution acknowledgment.
- Enhanced tests for `Desk`, `Balance`, and `Position` to validate new logic and ensure accurate behavior.
- Updated frontend state management to prevent overwriting non-empty snapshots in `positionsStore`.

+++ kraken/websocket/lifecycle.go
@@ -61,7 +61,7 @@ func (lifecycle *Lifecycle) Replay(trades []any) error {
 	return nil
 }
 
-func (lifecycle *Lifecycle) Place(model datura.Map[any], reqID int) error {
+func (lifecycle *Lifecycle) Place(model datura.Map[any], reqID int64) error {
 	orderAck := kraken.NewOrderResponseFromMap(model, reqID)
 
 	err := lifecycle.paper.simulator.Emit(lifecycle.paper, WEBSOCKET, "add_order", orderAck)
